Add ability for user to select role at start of game
This change adds new functionality to the CodeBreaker and CodeMaker classes. If the cpu is the code breaker, then sleep for one second before generating a random 4-digit integer guess for each turn. The sleep is to simulate the cpu 'thinking' before each guess. If the human is the code maker, then they will be able to create a secret code at the start of the game that the cpu will need to guess each turn